# docs/swagger.yaml
definitions:
...
  controller.BuiltinModelConfig:
    properties:
      cache_creation_price:
        type: number
      cached_price:
        type: number
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      exclude_from_tests:
        type: boolean
      image_batch_size:
        type: integer
      image_prices:
        additionalProperties:
          type: number
        type: object
      input_price:
        type: number
      model:
        type: string
      output_price:
        type: number
      owner:
        $ref: '#/definitions/model.ModelOwner'
      rpm:
        type: integer
      tpm:
        type: integer
      type:
        $ref: '#/definitions/relaymode.Mode'
      updated_at:
        type: string
    type: object
...
  controller.TestResult:
    properties:
      data:
        $ref: '#/definitions/model.ChannelTest'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  middleware.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  model.ChannelTest:
    properties:
      actual_model:
        type: string
      channel_id:
        type: integer
      channel_name:
        type: string
      channel_type:
        type: integer
      code:
        type: integer
      mode:
        $ref: '#/definitions/relaymode.Mode'
      model:
        type: string
      response:
        type: string
      success:
        type: boolean
      test_at:
        type: string
      took:
        type: number
    type: object
...
  model.ModelConfig:
    properties:
      cache_creation_price:
        type: number
      cached_price:
        type: number
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      exclude_from_tests:
        type: boolean
      image_batch_size:
        type: integer
      image_prices:
        additionalProperties:
          type: number
        type: object
      input_price:
        type: number
      model:
        type: string
      output_price:
        type: number
      owner:
        $ref: '#/definitions/model.ModelOwner'
      rpm:
        type: integer
      tpm:
        type: integer
      type:
        $ref: '#/definitions/relaymode.Mode'
      updated_at:
        type: string
    type: object
...
  model.ModelOwner:
    enum:
    - openai
    - alibaba
    - tencent
    - xunfei
    - deepseek
    - moonshot
    - minimax
    - baidu
    - google
    - baai
    - funaudiollm
    - doubao
    - fishaudio
    - chatglm
    - stabilityai
    - netease
    - ai360
    - anthropic
    - meta
    - baichuan
    - mistral
    - openchat
    - microsoft
    - defog
    - nexusflow
    - cohere
    - huggingface
    - lingyiwanwu
    - stepfun
    - xai
    - doc2x
    type: string
    x-enum-varnames:
    - ModelOwnerOpenAI
    - ModelOwnerAlibaba
    - ModelOwnerTencent
    - ModelOwnerXunfei
    - ModelOwnerDeepSeek
    - ModelOwnerMoonshot
    - ModelOwnerMiniMax
    - ModelOwnerBaidu
    - ModelOwnerGoogle
    - ModelOwnerBAAI
    - ModelOwnerFunAudioLLM
    - ModelOwnerDoubao
    - ModelOwnerFishAudio
    - ModelOwnerChatGLM
    - ModelOwnerStabilityAI
    - ModelOwnerNetease
    - ModelOwnerAI360
    - ModelOwnerAnthropic
    - ModelOwnerMeta
    - ModelOwnerBaichuan
    - ModelOwnerMistral
    - ModelOwnerOpenChat
    - ModelOwnerMicrosoft
    - ModelOwnerDefog
    - ModelOwnerNexusFlow
    - ModelOwnerCohere
    - ModelOwnerHuggingFace
    - ModelOwnerLingyiWanwu
    - ModelOwnerStepFun
    - ModelOwnerXAI
    - ModelOwnerDoc2x
...
  relaymode.Mode:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    type: integer
    x-enum-varnames:
    - Unknown
    - ChatCompletions
    - Completions
    - Embeddings
    - Moderations
    - ImagesGenerations
    - Edits
    - AudioSpeech
    - AudioTranscription
    - AudioTranslation
    - Rerank
    - ParsePdf
info:
  contact: {}
paths:
...
  /api/channel/{id}/models:
    get:
      description: Tests all models in the channel
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controller.TestResult'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Test channel models
      tags:
      - channel
...
  /api/dashboard/{group}/models:
    get:
      description: Returns model-specific metrics and usage data for the given group
      parameters:
      - description: Group name or ID
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ModelConfig'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get model usage data for a specific group
      tags:
      - dashboard
...
  /api/log/{group}/used/models:
    get:
      description: Get a list of models that have been used in a specific group's
        logs
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get group used models
      tags:
      - log
...
  /api/logs/used/models:
    get:
      description: Get a list of models that have been used in logs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get used models
      tags:
      - logs
...
  /api/models/builtin:
    get:
      description: Returns a list of builtin models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controller.BuiltinModelConfig'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get builtin models
      tags:
      - model
  /api/models/builtin/channel:
    get:
      description: Returns a list of channel builtin models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    items:
                      $ref: '#/definitions/controller.BuiltinModelConfig'
                    type: array
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channel builtin models
      tags:
      - model
  /api/models/builtin/channel/{type}:
    get:
      description: Returns a list of channel builtin models by type
      parameters:
      - description: Channel type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controller.BuiltinModelConfig'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channel builtin models by type
      tags:
      - model
  /api/models/default:
    get:
      description: Returns a list of channel default models and mapping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        mapping:
                          additionalProperties:
                            type: string
                          type: object
                        models:
                          items:
                            type: string
                          type: array
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channel default models and mapping
      tags:
      - model
  /api/models/default/{type}:
    get:
      description: Returns a list of channel default models and mapping by type
      parameters:
      - description: Channel type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        mapping:
                          additionalProperties:
                            type: string
                          type: object
                        models:
                          items:
                            type: string
                          type: array
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channel default models and mapping by type
      tags:
      - model
  /api/models/enabled:
    get:
      description: Returns a list of enabled models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ModelConfig'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get enabled models
      tags:
      - model
  /api/models/enabled/channel:
    get:
      description: Returns a list of channel enabled models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    items:
                      $ref: '#/definitions/model.ModelConfig'
                    type: array
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channel enabled models
      tags:
      - model
  /api/models/enabled/channel/{type}:
    get:
      description: Returns a list of channel enabled models by type
      parameters:
      - description: Channel type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ModelConfig'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channel enabled models by type
      tags:
      - model
...
  /api/monitor/models:
    get:
      description: Returns a list of models error rate
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    type: number
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get models error rate
      tags:
      - monitor
...
swagger: "2.0"